requirements.txt
praw
requests
tenacity
pandas
numpy
scikit-learn
networkx
python-dateutil
toml
tldextract
langid
matplotlib